Extract src-tauri/src/external.rs — sessions started outside Episko
One contiguous block: ExternalSession, ProcTable, parse_registry_entry,
list_external_sessions, owning_terminal and both arms of focus_external_session,
plus its 4 tests and their `table` fixture. lib.rs 1,940 -> 1,622.

The section divider became the module doc, and the imports sit under it rather
than above: the prose explains why LISTING is cross-platform while only focus is
macOS-only, which is the first thing a reader of this file needs.

sys_command turned out to be unused here — the focus path spawns `osascript`
directly — and ps_one is reached only from the macOS-only arms, so its import is
cfg-gated too.

Found by a new check, and it is worth keeping: `cargo check` on Windows never
compiles the cfg(not(windows)) arms, so a stale import in one of them is
invisible locally and a warning in CI. Flipping every cfg(windows) <->
cfg(not(windows)) on a scratch copy and re-checking type-checks the other half.
It reported zero errors and two warnings, one of them real:

- **the previous slice left a dead import**: lib.rs still had
  `#[cfg(not(windows))] use crate::platform::sh_quote;` after sh_quote's only
  consumer (spawn_external_terminal) moved to pty.rs. Removed here.
- `exists` unused in reveal_path is a flip artifact, not a finding: that function
  has three target_os arms and the flip doesn't touch cfg(target_os = ...), so on
  Windows the flip excluded all three.

Verification:
- external.rs vs the pre-slice lib.rs, line-sorted: 9 differences, all of them
  the rewritten section header and the imports I added.
- the remainder of lib.rs: 10 lines — mod external;, the two external::
  qualifications, and the import trim (git_repo_info and norm_path lost their
  last lib.rs user, plus the dead sh_quote above).

cargo test: 69 passed, same as baseline. cargo check --tests clean, no warnings,
and clean under the cfg flip too.
